wip: bring go module to work ...
